+++ storyf/story copy 7.html
@@ -1,0 +1,111 @@
+<!DOCTYPE html>
+<html lang="en"> 
+ <head>
+  <title>Eseis - About the Creators</title>
+  <meta name="keywords" content="choose your own adventure, story, hidden secrets, mystery"> <!--the keywords the search engine looks for when a person searches something-->
+  <meta charset="UTF-8">
+  <meta name="description" content="Follow Vesper as they discover the mystery behind their mother">
+  <meta name="author" content="Harshita Bhardwaj and Gerald Kyro Tungol"> <!--Author(s) of the website-->
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+   <!-- CSS only -->
+   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">  
+   
+   <!--Our CSS-->
+   <link href="../css/about_style.css" rel="stylesheet" type="text/css" />
+   <link href="../css/style.css" rel="stylesheet" type="text/css" />
+
+ </head>
+
+ <body>
+   
+    <!--NAV BAR-->
+    <nav class="navbar navbar-expand-sm navbar-dark bg-dark">
+      <div class="container">
+        <a 
+          href="../index.html"
+          class="navbar-brand mb-0 h1">
+
+          <!--Book Icon-->
+          <svg 
+            xmlns="http://www.w3.org/2000/svg" 
+            class="d-inline-block align-top bi bi-book-half" 
+            width="30" height="30" 
+            fill="currentColor" 
+            viewBox="0 0 16 16"
+            id="bookIcon">
+            <path d="M8.5 2.687c.654-.689 1.782-.886 3.112-.752 1.234.124 2.503.523 3.388.893v9.923c-.918-.35-2.107-.692-3.287-.81-1.094-.111-2.278-.039-3.213.492V2.687zM8 1.783C7.015.936 5.587.81 4.287.94c-1.514.153-3.042.672-3.994 1.105A.5.5 0 0 0 0 2.5v11a.5.5 0 0 0 .707.455c.882-.4 2.303-.881 3.68-1.02 1.409-.142 2.59.087 3.223.877a.5.5 0 0 0 .78 0c.633-.79 1.814-1.019 3.222-.877 1.378.139 2.8.62 3.681 1.02A.5.5 0 0 0 16 13.5v-11a.5.5 0 0 0-.293-.455c-.952-.433-2.48-.952-3.994-1.105C10.413.809 8.985.936 8 1.783z"/>
+          </svg>
+
+          <!--Website Name-->
+          Eseis
+        </a>
+
+        <!--Collapse button and the buttons that will be collapsed-->
+        <button 
+          type="button"
+          data-bs-toggle="collapse"
+          data-bs-target="#navbarNav"
+          class="navbar-toggler"
+          aria-controls="navbarNav"
+          aria-expanded="false"
+          aria-label="Toggle Navigation">
+
+          <span class="navbar-toggler-icon"></span> <!--Icon for the button-->
+        </button>
+        <div 
+          class="collapse navbar-collapse"
+          id="navbarNav">
+
+          <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
+            <li class="nav-item">
+              <a href = "../index.html" class="nav-link">
+                Home
+              </a>
+            </li>
+
+            <li class="nav-item">
+              <a href = "../pages/lore.html" class="nav-link">
+                Lore
+              </a>
+            </li>
+
+            <li class="nav-item active">
+              <a href = "../pages/aboutUs.html" class="nav-link active">
+                About Us
+              </a>
+            </li>
+
+          </ul>
+        </div>
+      </div>
+    </nav>
+    
+    <div class="card container border-0 harshita" style="width: 100%">
+      <img class="card-img rounded-circle harshita" src="../images/websiteH.jpg" alt="Harshita Bhardwaj Picture">
+      <div class="card-body harshita">
+        <h5 class="card-title harshita">Harshita Bhardwaj</h5>
+        <p class="card-text harshita">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
+      </div>
+    </div>
+    <div class="card text-end container-fluid border-0 gerald" style="width: 100%">
+      <div class="card-body gerald">
+        <h5 class="card-title gerald">Gerald Kyro Tungol</h5>
+        <p class="card-text gerald">Gerald Kyro Tungol pursuing a career in this general field. On his path of education he is prepared to learn and develop new skills required for the projects he will participate in. Games! 🎮. </p>
+      </div>
+      <img class="card-img rounded-circle" src="../images/websiteG.jpg" alt="Gerald Kyro Tungol Picture">
+    </div>
+
+    <div class="footer mt-5 p-3"></div> <!--Just a space (Maybe add an easter egg here)-->
+    
+
+    <!-- JavaScript Bundle with Popper -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
+
+    <!--Our JS-->
+    <script src="./js/script.js"></script>
+
+
+  </body>
+</html>
+
